@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { FormControl, Validators } from '@angular/forms';
+import { UntypedFormControl, Validators } from '@angular/forms';
 import { Router } from '@angular/router';
 import { ToastrService } from 'ngx-toastr';
 import { Credenciais } from 'src/app/models/credenciais';
@@ -18,8 +18,8 @@ export class LoginComponent implements OnInit {
   }
 
   hide = true;
-  email = new FormControl(null,Validators.email);
-  senha = new FormControl(null,Validators.minLength(3));
+  email = new UntypedFormControl(null,Validators.email);
+  senha = new UntypedFormControl(null,Validators.minLength(3));
 
   constructor(private toast: ToastrService, private service: AuthService, private router: Router){}
 
